fix some wrong 'double' to 'int' assignments

+++ src/spicelib/devices/csw/cswacld.c
@@ -33,7 +33,7 @@ CSWacLoad(GENmodel *inModel, CKTcircuit *ckt)
                 here=here->CSWnextInstance) {
 	    if (here->CSWowner != ARCHme) continue;
 
-            current_state = *(ckt->CKTstate0 + here->CSWstate);
+            current_state = (int)*(ckt->CKTstate0 + here->CSWstate);
 
             g_now = current_state?(model->CSWonConduct):(model->CSWoffConduct);
